[Succeed]ccache: stable 3.3.4 (bottled), HEAD Object-file caching compiler wrapper https://ccache.samba.org/ Not installed From: https://github.com/Linuxbrew/homebrew-core/blob/master/Formula/ccache.rb ==> Dependencies Build: xz ==> Caveats To install symlinks for compilers that will automatically use ccache, prepend this directory to your PATH:   /home/linuxbrew/.linuxbrew/opt/ccache/libexec
If this is an upgrade and you have previously added the symlinks to
your PATH, you may need to modify it to the path specified above so
it points to the current version.

NOTE: ccache can prevent some software from compiling.
ALSO NOTE: The brew command, by design, will never use ccache.
4.4.0-1039-aws #48-Ubuntu SMP Wed Oct 11 15:15:01 UTC 2017 GNU/Linux